fixes infinite sleep due to uint32_t timestamps

+++ src/main.cpp
@@ -111,7 +111,7 @@ int main(int argc, char * argv[])
         // Getting time difference
         gettimeofday(&tv2, NULL);
         elapsed = tv2.tv_sec * 1000000 + tv2.tv_usec - (tv1.tv_sec * 1000000 + tv1.tv_usec);
-        if (period - elapsed > 0) {
+        if ((int) period - (int) elapsed > 0) {
             usleep(period - elapsed);
         }
     }
